refactor: 전송 경계와 파일 액션 오류 처리를 정리
- 브라우저 전송 오케스트레이션을 도메인별 훅과 핸들러로 분리한다\n- Space root 사전 검증과 browse 권한 안내를 일관되게 맞춘다\n- 파일 액션 오류 분류와 zip 실패 전파를 보강한다\n\nRelated: #252\nCloses: #253
